feat(entidades): Se crearon todas las tablas faltantes.

+++ hotel-premier/src/main/java/com/losmergeconflicts/hotelpremier/entity/ItemConsumo.java
@@ -26,5 +26,5 @@ public class ItemConsumo {
     private String nombre;
 
     @Column(nullable = false)
-    private Double precio;
+    private Float precio;
 }
